+++ examples/status/main.go
@@ -3,14 +3,17 @@ package main
 import (
     "fmt"
 
-    "github.com/mondegor/go-components/mrcom"
+    mrcom_status "github.com/mondegor/go-components/mrcom/status"
 )
 
 func main() {
-    status := mrcom.ItemStatusEnabled
+    status := mrcom_status.ItemStatusEnabled
 
     fmt.Printf("STATUS: %s\n", status.String())
 
-    fmt.Printf("check: %#v\n", mrcom.ItemStatusFlowDefault.Check(mrcom.ItemStatusEnabled, mrcom.ItemStatusDisabled))
-    fmt.Printf("check: %#v\n", mrcom.ItemStatusFlowDefault.Check(mrcom.ItemStatusRemoved, mrcom.ItemStatusDisabled))
+    fmt.Printf("check: %#v\n", mrcom_status.ItemStatusFlow.Check(mrcom_status.ItemStatusEnabled, mrcom_status.ItemStatusDisabled))
+    fmt.Printf("check: %#v\n", mrcom_status.ItemStatusFlow.Check(mrcom_status.ItemStatusRemoved, mrcom_status.ItemStatusDisabled))
+
+    fmt.Printf("check: %#v\n", mrcom_status.OnlyRemoveStatusFlow.Check(mrcom_status.OnlyRemoveStatusEnabled, mrcom_status.OnlyRemoveStatusRemoved))
+    fmt.Printf("check: %#v\n", mrcom_status.OnlyRemoveStatusFlow.Check(mrcom_status.OnlyRemoveStatusRemoved, mrcom_status.OnlyRemoveStatusEnabled))
 }